pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0", "numpy >= 2.0.0"]
build-backend = "maturin"

[tool.pytest.ini_options]
minversion = "6.2.2"
addopts = [
    "--import-mode=importlib",
    "--benchmark-skip",
    "--benchmark-group-by=group",
    "--benchmark-warmup=on",
    "--benchmark-warmup-iterations=5",
    "--benchmark-columns=min,max,mean,stddev,ops",
    "--benchmark-histogram=benchmark_results",
]
testpaths = [
    "tests",
]

[project]
name = "ckmeans"
dynamic = ["version"]
description = "Optimal univariate (1D) clustering based on Ckmeans.1d.dp"
authors = [{ name = "Stephan Hügel", email = "urschrei@gmail.com" }]
license = {file = "license.txt"}
keywords = ["ckmeans", "clustering", "jenks"]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Blue Oak Model License (BlueOak-1.0.0)",
    "Operating System :: OS Independent",
]
[project.urls]
Repository = "https://github.com/urschrei/ckmeans_py"
Tracker = "https://github.com/urschrei/ckmeans_py/issues"
[tool.maturin]
features = ["pyo3/extension-module"]

[dependency-groups]
dev = [
    "ckmeans-1d-dp>=4.3.4.4",
    "pytest-benchmark[histogram]>=5.1.0",
]
[tool.maturin.dynamic]
